centralise data CTFE result extraction layer:
  New file: src/codegen/type_marshal.d
  - MarshalledValue — tagged union result type with kinds: integer, float_, slice, staticArray, struct_, raw
  - TypeReader — target-parameterized struct with forWasm() and forNative() factory methods
    - readInt(buf, size) — signed integer (1/2/4/8 bytes)
    - readUint(buf, size) — unsigned integer
    - readFloat(buf, size) — float/double
    - readPtr(buf) — pointer (4 or 8 bytes depending on target)
    - readSlice(buf) — slice header {ptr, len, cap} with proper target offsets
    - readValue(buf, type) — dispatches to the right reader based on D type
    - readStruct(buf, decl) — reads struct fields using layout info
    - elementSizeOf(type) — static method to compute element size
  - All methods have pre-condition asserts before pointer casts
  - Inline unit tests covering all readers for both WASM and native layouts

  New file: tests/test_type_marshal.d
  - Dedicated test module with separate unittest blocks for each method

  Modified: src/codegen/backend.d
  - CompiledFunction.readMemory(uint offset, ...) → readMemory(ulong offset, ...)

  Modified: src/codegen/wasm/backend.d
  - readMemory signature updated; casts ulong to uint for WASM linear memory

  Modified: src/codegen/native/backend.d
  - readMemory signature updated; now handles 64-bit native pointers correctly
  - Added null pointer assert

  Modified: src/semantic/ctfe.d
  - Added import codegen.type_marshal
  - Added TypeReader reader field on CTFEEvaluator
  - Initialized from backendName in constructor
  - String return path: replaced hardcoded 12-byte slice extraction with reader.readSlice()
  - Dynamic array return path: replaced hardcoded extraction with reader.readSlice() +
  TypeReader.elementSizeOf()

  The key win: one code path for both backends — no more if (isNative) branches or hardcoded WASM offsets. The
   TypeReader handles the layout differences via its target-specific constants.

+++ src/codegen/backend.d
@@ -65,7 +65,7 @@ interface CompiledFunction {
     ExecutionResult callWithLargeReturn(string funcName, long[] args, uint resultSize);
     
     /// Read bytes from the execution memory (WASM linear memory or native address space)
-    ubyte[] readMemory(uint offset, uint length);
+    ubyte[] readMemory(ulong offset, uint length);
 
     /// Check if a function exists in this compiled context
     bool hasFunction(string funcName);
